UI
1. TodoTab.js에 있던 AddTagWindow를 App.js로 이동
2. TodoStore.js에 editTag시, 기존에 설정되어있는 태그들이 바뀌지 않는 버그가 있어 재접속시 해결되지만 즉시 반영하기 위해 todo state 재설정
3. App.js 등에 불필요한 부분 삭제
