+++ lib/TodoStore.js
@@ -99,9 +99,14 @@ export function useTodo() {
   }
 
   function editTag(i, tag) {
+    const filterTodo = todoState.map((todo) => {
+      todo.tags = todo.tags.map((v) => (v === tagState[i] ? tag : v));
+      return todo;
+    });
     let copy = tagState.slice();
     copy.splice(i, 1, { ...tag, id: tagState[i].id });
     setTagState(copy);
+    setTodoState(filterTodo);
   }
 
   function removeTag(i) {
